+++ src/tls.cpp
@@ -1,13 +1,13 @@
 #include "tls.h"
 
-int16_t TLS::init() {
+int TLS::init() {
 	return 0;
 }
 
-int16_t TLS::connect(const char *host, const char *port) {
+int TLS::connect(const char *host, const char *port) {
 	return mbedtls_net_connect( &server_fd, host, port, MBEDTLS_NET_PROTO_TCP );
 }
 
-int16_t TLS::write(const char *buf) {
+int TLS::write(const char *buf) {
 	return MBEDTLS_ERR_SSL_WANT_READ;
 }
